Rework embedding provider module structure
