+++ crates/rpc/src/proto.rs
@@ -165,6 +165,8 @@ messages!(
     (DeleteNotification, Foreground),
     (UpdateNotification, Foreground),
     (DeleteProjectEntry, Foreground),
+    (EmbedTexts, Background),
+    (EmbedTextsResponse, Background),
     (EndStream, Foreground),
     (Error, Foreground),
     (ExpandProjectEntry, Foreground),
@@ -320,6 +322,7 @@ request_messages!(
     (DeclineCall, Ack),
     (DeleteChannel, Ack),
     (DeleteProjectEntry, ProjectEntryResponse),
+    (EmbedTexts, EmbedTextsResponse),
     (ExpandProjectEntry, ExpandProjectEntryResponse),
     (Follow, FollowResponse),
     (FormatBuffers, FormatBuffersResponse),
