+++ crates/semantic_index/src/embedding.rs
@@ -1,25 +1,15 @@
-use anyhow::{anyhow, Context as _, Result};
-use futures::{future::BoxFuture, AsyncReadExt, FutureExt};
+mod ollama;
+mod openai;
+mod zed;
+
+pub use ollama::*;
+pub use openai::*;
+pub use zed::*;
+
+use anyhow::Result;
+use futures::{future::BoxFuture, FutureExt};
 use serde::{Deserialize, Serialize};
-use std::{fmt, future, sync::Arc};
-use util::http::{AsyncBody, HttpClient, Method, Request as HttpRequest};
-
-/// Ollama's embedding via nomic-embed-text is of length 768
-pub const EMBEDDING_SIZE_TINY: usize = 768;
-/// Ollama's embedding via mxbai-embed-large is of length 1024
-pub const EMBEDDING_SIZE_XSMALL: usize = 1024;
-/// OpenAI's text small and Voyage Large/Code are of length 1536
-pub const EMBEDDING_SIZE_SMALL: usize = 1536;
-/// OpenAI's text large embeddings are of length 3072
-pub const EMBEDDING_SIZE_LARGE: usize = 3072;
-
-#[derive(Clone, Copy)]
-pub enum EmbeddingModel {
-    OllamaNomicEmbedText,
-    OllamaMxbaiEmbedLarge,
-    OpenaiTextEmbedding3Small,
-    OpenaiTextEmbedding3Large,
-}
+use std::{fmt, future};
 
 #[derive(Debug, Default, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Embedding(Vec<f32>);
@@ -83,20 +73,14 @@ pub trait EmbeddingProvider: Sync + Send {
     fn batch_size(&self) -> usize;
 }
 
-pub struct FakeEmbeddingProvider {}
-
-impl FakeEmbeddingProvider {
-    pub fn new() -> Self {
-        Self {}
-    }
-}
+pub struct FakeEmbeddingProvider;
 
 impl EmbeddingProvider for FakeEmbeddingProvider {
     fn embed(&self, texts: &[&str]) -> BoxFuture<Result<Vec<Embedding>>> {
         let embeddings = texts
             .iter()
             .map(|text| {
-                let mut embedding = vec![0f32; 4];
+                let mut embedding = vec![0f32; 1536];
                 for i in 0..embedding.len() {
                     embedding[i] = i as f32;
                 }
@@ -111,239 +95,9 @@ impl EmbeddingProvider for FakeEmbeddingProvider {
     }
 }
 
-pub struct OllamaEmbeddingProvider {
-    client: Arc<dyn HttpClient>,
-    model: EmbeddingModel,
-}
-
-#[derive(Serialize)]
-struct OllamaEmbeddingRequest {
-    model: String,
-    prompt: String,
-}
-
-#[derive(Deserialize)]
-struct OllamaEmbeddingResponse {
-    embedding: Vec<f32>,
-}
-
-impl OllamaEmbeddingProvider {
-    pub fn new(client: Arc<dyn HttpClient>, model: EmbeddingModel) -> Self {
-        Self { client, model }
-    }
-}
-
-impl EmbeddingProvider for OllamaEmbeddingProvider {
-    fn embed(&self, texts: &[&str]) -> BoxFuture<Result<Vec<Embedding>>> {
-        //
-        let model = match self.model {
-            EmbeddingModel::OllamaNomicEmbedText => "nomic-embed-text",
-            EmbeddingModel::OllamaMxbaiEmbedLarge => "mxbai-embed-large",
-            _ => return future::ready(Err(anyhow!("Invalid model"))).boxed(),
-        };
-
-        futures::future::try_join_all(texts.into_iter().map(|text| {
-            let request = OllamaEmbeddingRequest {
-                model: model.to_string(),
-                prompt: text.to_string(),
-            };
-
-            let request = serde_json::to_string(&request).unwrap();
-
-            async {
-                let response = self
-                    .client
-                    .post_json("http://localhost:11434/api/embeddings", request.into())
-                    .await?;
-
-                let mut body = String::new();
-                response.into_body().read_to_string(&mut body).await?;
-
-                let response: OllamaEmbeddingResponse =
-                    serde_json::from_str(&body).context("Unable to pull response")?;
-
-                Ok(Embedding::new(response.embedding))
-            }
-        }))
-        .boxed()
-    }
-
-    fn batch_size(&self) -> usize {
-        // TODO: Figure out decent value
-        10
-    }
-}
-
-pub struct OpenaiEmbeddingProvider {
-    client: Arc<dyn HttpClient>,
-    model: EmbeddingModel,
-    api_key: String,
-}
-
-#[derive(Serialize)]
-struct OpenaiEmbeddingRequest {
-    model: String,
-    input: Vec<String>,
-}
-
-#[derive(Deserialize)]
-struct OpenaiEmbeddingData {
-    embedding: Vec<f32>,
-}
-
-#[derive(Deserialize)]
-struct OpenaiEmbeddingResponse {
-    object: String,
-    data: Vec<OpenaiEmbeddingData>,
-    model: String,
-}
-
-impl OpenaiEmbeddingProvider {
-    pub fn new(client: Arc<dyn HttpClient>, model: EmbeddingModel, api_key: String) -> Self {
-        Self {
-            client,
-            model,
-            api_key,
-        }
-    }
-}
-
-impl EmbeddingProvider for OpenaiEmbeddingProvider {
-    fn embed(&self, texts: &[&str]) -> BoxFuture<Result<Vec<Embedding>>> {
-        let model = match self.model {
-            EmbeddingModel::OpenaiTextEmbedding3Small => "text-embedding-3-small",
-            EmbeddingModel::OpenaiTextEmbedding3Large => "text-embedding-3-large",
-            _ => return future::ready(Err(anyhow!("Invalid model"))).boxed(),
-        };
-
-        // Unlike the Ollama model, we can send `texts` as a batch directly
-
-        let api_url = "https://api.openai.com/v1/";
-
-        let uri = format!("{api_url}/embeddings");
-
-        let request = OpenaiEmbeddingRequest {
-            model: model.to_string(),
-            input: texts.iter().map(|text| text.to_string()).collect(),
-        };
-        let request = serde_json::to_string(&request).unwrap();
-        let body = AsyncBody::from(request);
-
-        let request = HttpRequest::builder()
-            .method(Method::POST)
-            .uri(uri)
-            .header("Content-Type", "application/json")
-            .header("Authorization", format!("Bearer {}", self.api_key))
-            .body(body);
-
-        let request = if let Ok(request) = request {
-            request
-        } else {
-            return future::ready(Err(anyhow!("Failed to build request"))).boxed();
-        };
-
-        async {
-            let response = self.client.send(request).await?;
-
-            let mut body = String::new();
-            response.into_body().read_to_string(&mut body).await?;
-
-            // todo(): check for errors in response (likely due to rate limiting or invalid API key)
-            let response: OpenaiEmbeddingResponse = serde_json::from_str(&body)
-                .context("Response format from OpenAI did not match struct")?;
-
-            let embeddings = response
-                .data
-                .into_iter()
-                .map(|data| Embedding::new(data.embedding))
-                .collect();
-
-            Ok(embeddings)
-        }
-        .boxed()
-    }
-
-    fn batch_size(&self) -> usize {
-        // From https://platform.openai.com/docs/api-reference/embeddings/create
-        2048
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
-    use gpui::BackgroundExecutor;
-
-    use util::http::HttpClientWithUrl;
-
-    #[gpui::test]
-    async fn test_ollama_embedding_provider(executor: BackgroundExecutor) {
-        executor.allow_parking();
-
-        let client = Arc::new(HttpClientWithUrl::new("http://localhost:11434/"));
-        let provider =
-            OllamaEmbeddingProvider::new(client.clone(), EmbeddingModel::OllamaNomicEmbedText);
-        let embeddings = provider.embed(&[&"Hello, world!"]).await.unwrap();
-        assert_eq!(embeddings[0].len(), EMBEDDING_SIZE_TINY);
-    }
-
-    #[gpui::test]
-    async fn test_ollama_embedding_not_exactly_a_benchmark(executor: BackgroundExecutor) {
-        executor.allow_parking();
-
-        let client = Arc::new(HttpClientWithUrl::new("http://localhost:11434/"));
-        let provider =
-            OllamaEmbeddingProvider::new(client.clone(), EmbeddingModel::OllamaNomicEmbedText);
-
-        let t_nomic = std::time::Instant::now();
-        let strings: Vec<String> = (0..100).map(|i| format!("Hello, world! {}", i)).collect();
-        let texts: Vec<&str> = strings.iter().map(AsRef::as_ref).collect();
-
-        let embeddings = provider.embed(&texts).await.unwrap();
-        for embedding in embeddings {
-            assert_eq!(embedding.len(), EMBEDDING_SIZE_TINY);
-        }
-        dbg!(t_nomic.elapsed());
-
-        let client = Arc::new(HttpClientWithUrl::new("http://localhost:11434/"));
-        let provider =
-            OllamaEmbeddingProvider::new(client.clone(), EmbeddingModel::OllamaMxbaiEmbedLarge);
-
-        let t_mxbai = std::time::Instant::now();
-
-        let strings: Vec<String> = (0..100).map(|i| format!("Hello, world! {}", i)).collect();
-        let texts: Vec<&str> = strings.iter().map(AsRef::as_ref).collect();
-
-        let embeddings = provider.embed(&texts).await.unwrap();
-        for embedding in embeddings {
-            assert_eq!(embedding.len(), EMBEDDING_SIZE_XSMALL);
-        }
-        dbg!(t_mxbai.elapsed());
-    }
-
-    #[gpui::test]
-    async fn test_openai_embedding(executor: BackgroundExecutor) {
-        executor.allow_parking();
-
-        let api_key = std::env::var("OPENAI_API_KEY").expect("OPENAI_API_KEY not set");
-
-        let client = Arc::new(HttpClientWithUrl::new("http://localhost:11434/"));
-        let provider = OpenaiEmbeddingProvider::new(
-            client.clone(),
-            EmbeddingModel::OpenaiTextEmbedding3Small,
-            api_key,
-        );
-
-        let strings: Vec<String> = (0..100).map(|i| format!("Hello, world! {}", i)).collect();
-        let texts: Vec<&str> = strings.iter().map(AsRef::as_ref).collect();
-
-        let t_openai_small = std::time::Instant::now();
-        let embeddings = provider.embed(&texts).await.unwrap();
-        for embedding in embeddings {
-            assert_eq!(embedding.len(), EMBEDDING_SIZE_SMALL);
-        }
-        dbg!(t_openai_small.elapsed());
-    }
 
     #[gpui::test]
     fn test_normalize_embedding() {
